Remove private data/scripts and update dependencies
Commit is removing Docker and networking scripts that are no longer needed due to changes in the deployment model. This includes startDB.sh, startAPP.sh, build.sh, clear_docker.sh, db_app_network.sh, and .env and altering .gitignore to include deleted files. It also updates the versions of the project's dependencies in requirements.txt. The changes aim to keep the repository clean and up-to-date.